chore: remove dead code and rename transformers to safe-logging
Removed 607 lines of unused utility functions and test code from the
config transformers module, then renamed it to reflect its single
purpose.

Changes:
- Remove dead code: validateConfig, normalizePort, sanitizeAlgorithmList,
  normalizeAlgorithms, validateCorsOrigins, createClientSafeConfig,
  validateSessionConfig, mergeMultipleConfigs (never called in production)
- Remove 27 tests for dead functions (285 lines)
- Remove mergeConfigs and applyDefaults wrappers (duplicates removed earlier)
- Rename transformers.ts → safe-logging.ts for clarity
- Update import in app/config.ts
- Rename test file: transformers.vitest.ts → safe-logging.vitest.ts

Result:
- app/config/safe-logging.ts: 97 lines (was 308, -68%)
- tests: 5 tests (was 32, kept only production-critical tests)
- All 595 tests pass
- Lint and typecheck clean
- Single clear purpose: mask sensitive config data for safe logging

The module now has a single export (maskSensitiveConfig) used once in
production code (config.ts:147) to prevent secrets from appearing in
debug logs.
